requirements.txt
asgiref==3.2.3
backports.entry-points-selectable==1.1.1
distlib==0.3.3
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
filelock==3.4.0
mysqlclient==2.1.0
Pillow==8.4.0
platformdirs==2.4.0
pytz==2019.3
six==1.16.0
sqlparse==0.3.0
virtualenv==20.10.0
